Restructure for Classic/Modern split (#74)
This does some basic restructuring to the mutation
notation modules and cleans up the macro logic. It also:

- Adds the deprecation warning for defaulting behavior
- Add some additional examples, docs
