--- lib/absinthe/relay/mutation/modern_notation.ex
@@ -1,9 +1,0 @@
-defmodule Absinthe.Relay.Mutation.ModernNotation do
-  use Absinthe.Relay.Mutation.Base
-
-  defp client_mutation_id_field do
-    quote do
-      field :client_mutation_id, type: :string
-    end
-  end
-end
