+++ lib/absinthe/relay/mutation/notation/classic.ex
@@ -1,0 +1,10 @@
+defmodule Absinthe.Relay.Mutation.Notation.Classic do
+  @moduledoc """
+  Macros used to define Mutation-related schema entities for Relay Classic.
+
+  See `Absinthe.Relay.Mutation` for examples of use.
+  """
+
+  use Absinthe.Relay.Mutation.Notation
+
+end
